+++ src/simulation/viz_interface/vizInterface.cpp
@@ -119,7 +119,7 @@ void VizInterface::SelfInit()
         
         /*! - Create output message for module in opNav mode */
         if (this->opNavMode > 0) {
-            int imageBufferCount = 2;
+            uint64_t imageBufferCount = 2;
             this->imageOutMsgID = SystemMessaging::GetInstance()->CreateNewMessage(this->opnavImageOutMsgName,sizeof(CameraImageMsg),imageBufferCount,"CameraImageMsg", moduleID);
         }
     }
@@ -179,7 +179,7 @@ void VizInterface::CrossInit()
 
     /*! Define Spice input message */
     {
-        int i=0;
+        uint i=0;
         MsgCurrStatus spiceStatus;
         spiceStatus.dataFresh = false;
         spiceStatus.lastTimeTag = 0xFFFFFFFFFFFFFFFF;
@@ -216,7 +216,7 @@ void VizInterface::CrossInit()
         MsgCurrStatus rwStatus;
         rwStatus.dataFresh = false;
         rwStatus.lastTimeTag = 0xFFFFFFFFFFFFFFFF;
-        for (int idx = 0; idx < this->numRW; idx++)
+        for (uint idx = 0; idx < this->numRW; idx++)
         {
             std::string tmpWheelMsgName = "rw_config_" + std::to_string(idx) + "_data";
             this->rwInMsgName.push_back(tmpWheelMsgName);
@@ -292,7 +292,7 @@ void VizInterface::ReadBSKMessages()
     }
     /*! Read BSK Spice constellation msg */
     {
-    int i=0;
+    uint i=0;
     std::vector<std::string>::iterator it;
     for(it = this->planetNames.begin(); it != this->planetNames.end(); it++)
     {
@@ -314,7 +314,7 @@ void VizInterface::ReadBSKMessages()
 
     /*! Read BSK RW constellation msg */
     {
-    for (int idx=0;idx< this->numRW; idx++)
+    for (uint idx=0;idx< this->numRW; idx++)
     {
         if (this->rwInMsgID[idx].msgID != -1){
         RWConfigLogSimMsg localRWArray;
@@ -332,7 +332,7 @@ void VizInterface::ReadBSKMessages()
     
      /*! Read incoming Thruster constellation msg */
     {
-    for (int idx=0;idx< this->numThr; idx++){
+    for (uint idx=0;idx< this->numThr; idx++){
         if (this->thrMsgID[idx].msgID != -1){
             THROutputSimMsg localThrusterArray;
             SingleMessageHeader localThrusterHeader;
@@ -445,7 +445,7 @@ void VizInterface::WriteProtobuffer(uint64_t CurrentSimNanos)
         }
 
         // define any pointing lines for Vizard
-        for (int idx = 0; idx < this->settings.pointLineList.size(); idx++) {
+        for (uint idx = 0; idx < this->settings.pointLineList.size(); idx++) {
             vizProtobufferMessage::VizMessage::PointLine* pl = vizSettings->add_pointlines();
             pl->set_tobodyname(this->settings.pointLineList[idx].toBodyName);
             pl->set_frombodyname(this->settings.pointLineList[idx].fromBodyName);
@@ -455,7 +455,7 @@ void VizInterface::WriteProtobuffer(uint64_t CurrentSimNanos)
         }
 
         // define any keep in/out cones for Vizard
-        for (int idx = 0; idx < this->settings.coneList.size(); idx++) {
+        for (uint idx = 0; idx < this->settings.coneList.size(); idx++) {
             vizProtobufferMessage::VizMessage::KeepOutInCone* cone = vizSettings->add_keepoutincones();
             cone->set_iskeepin(this->settings.coneList[idx].isKeepIn);
             for (int i=0; i<3; i++) {
@@ -473,7 +473,7 @@ void VizInterface::WriteProtobuffer(uint64_t CurrentSimNanos)
         }
 
         // define actuator GUI settings
-        for (int idx = 0; idx < this->settings.actuatorGuiSettingsList.size(); idx++) {
+        for (uint idx = 0; idx < this->settings.actuatorGuiSettingsList.size(); idx++) {
             vizProtobufferMessage::VizMessage::ActuatorSettings* al = vizSettings->add_actuatorsettings();
             al->set_spacecraftname(this->settings.actuatorGuiSettingsList[idx].spacecraftName);
             al->set_viewthrusterpanel(this->settings.actuatorGuiSettingsList[idx].viewThrusterPanel);
@@ -534,7 +534,7 @@ void VizInterface::WriteProtobuffer(uint64_t CurrentSimNanos)
         //scPlusInMsgID.dataFresh = false;
 
         /*! Write RW output msg */
-        for (int idx =0; idx < this->numRW; idx++)
+        for (uint idx =0; idx < this->numRW; idx++)
         {
             if (rwInMsgID[idx].msgID != -1 && rwInMsgID[idx].dataFresh){
                 vizProtobufferMessage::VizMessage::ReactionWheel* rwheel = scp->add_reactionwheels();
@@ -549,7 +549,7 @@ void VizInterface::WriteProtobuffer(uint64_t CurrentSimNanos)
         }
 
         /*! Write Thr output msg */
-        for (int idx =0; idx < this->numThr; idx++)
+        for (uint idx =0; idx < this->numThr; idx++)
         {
             if (thrMsgID[idx].msgID != -1 && thrMsgID[idx].dataFresh){
                 vizProtobufferMessage::VizMessage::Thruster* thr = scp->add_thrusters();
@@ -616,7 +616,7 @@ void VizInterface::WriteProtobuffer(uint64_t CurrentSimNanos)
 
 
         /*! Write spice output msgs */
-        int k=0;
+        uint k=0;
         std::vector<std::string>::iterator it;
         for(it = this->planetNames.begin(); it != this->planetNames.end(); it++)
         {
@@ -640,9 +640,9 @@ void VizInterface::WriteProtobuffer(uint64_t CurrentSimNanos)
         google::protobuf::uint8 varIntBuffer[4];
         uint32_t byteCount = message->ByteSizeLong();
         google::protobuf::uint8 *end = google::protobuf::io::CodedOutputStream::WriteVarint32ToArray(byteCount, varIntBuffer);
-        unsigned long varIntBytes = end - varIntBuffer;
+        unsigned long varIntBytes = (unsigned long) (end - varIntBuffer);
         if (this->saveFile) {
-            this->outputStream->write(reinterpret_cast<char* > (varIntBuffer), varIntBytes);
+            this->outputStream->write(reinterpret_cast<char* > (varIntBuffer), (int) varIntBytes);
         }
 
         /*! Enter in lock-step with the vizard to simulate a camera */
@@ -663,7 +663,7 @@ void VizInterface::WriteProtobuffer(uint64_t CurrentSimNanos)
             
             /*! - send protobuffer raw over zmq_socket */
             void* serialized_message = malloc(byteCount);
-            message->SerializeToArray(serialized_message, byteCount);
+            message->SerializeToArray(serialized_message, (int) byteCount);
 
             /*! - Normal sim step by sending protobuffers */
             zmq_msg_t request_header;
